milestone2/utils/llm.py
from langchain_openai import ChatOpenAI
from dotenv import load_dotenv
import os

load_dotenv()

def get_llm():

    llm = ChatOpenAI(
        model="qwen2.5-7b-instruct-1m",
        base_url=os.getenv("OPENAI_BASE_URL"),
        api_key=os.getenv("OPENAI_API_KEY"),
        temperature=0.3
    )

    return llm